Exception Handling
use errorcontroller and custom error view(jsp)

+++ 02 Web Application with Spring Boot/first_app/src/main/java/com/example/demo/controller/TodoController.java
@@ -60,6 +60,9 @@ public class TodoController {
 
 	@RequestMapping(value = "/delete-todo", method = RequestMethod.GET)
 	public String deleteAddTodo(@RequestParam int id) {
+//		to show exception handling working
+		if (id == 1)
+			throw new RuntimeException("Something went wrong");
 		service.deleteTodo(id);
 		return "redirect:/list-todos";
 	}
